# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UptimePulse API
  version: 0.0.1
  description: |
    API REST de UptimePulse (monitorización de uptime). Se sirve en `/docs`
    (Swagger UI) desde la propia API.

    **Autenticación.** Dos formas, ambas por `Authorization: Bearer …`:
    - un *access token* JWT de 5 min (de `/auth/register`, `/auth/login` o
      `/auth/refresh`; el refresh token viaja en una cookie httpOnly y caduca
      a los `SESSION_IDLE_TIMEOUT_MINUTES` de inactividad, 15 por defecto,
      renovándose en cada refresh);
    - una **API key** `up_…` creada por un admin en
      `/organizations/{id}/api-keys`. Actúa en nombre de la organización con
      el rol equivalente a sus scopes (`read` → readonly, `write` → editor) y
      nunca puede administrar la organización (`/me`, equipo, claves).

    **Organización activa.** Las rutas de datos actúan sobre la organización
    indicada en `X-Organization-Id`; sin cabecera, sobre la personal del
    usuario. Con API key la fija la clave y la cabecera se ignora.

    **Errores.** Siempre `{ "error": "<mensaje en español>" }`; los 400 de
    validación pueden llevar el detalle de zod en `error`. Cada respuesta
    lleva `X-Request-Id`.

    **Rate limits.** 300 peticiones/min por IP o API key en toda la API (429);
    login y registro 10/min; `/public/status`, `/invitations/{token}`
    y `/auth/username-available` 60/min.
servers:
  - url: /api
    description: Self-hosted (detrás del nginx del frontend, mismo origen)
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: JWT de acceso o API key `up_…`.
  parameters:
    organizationHeader:
      name: X-Organization-Id
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: Organización activa. Sin ella, la personal del usuario.
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          description: Mensaje legible, o el detalle de validación de zod.
          oneOf: [{ type: string }, { type: object }]
...
    Incident:
      type: object
      properties:
        id: { type: string, format: uuid }
        monitorId: { type: string, format: uuid }
        startedAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time, nullable: true }
        causeSummary: { type: string, nullable: true }
...
    StatusPageInput:
      type: object
      required: [slug, title]
      properties:
        slug: { type: string, pattern: "^[a-z0-9-]+$", maxLength: 60, description: Único dentro de la organización. }
        title: { type: string, maxLength: 200 }
        isPublic: { type: boolean, default: true }
        monitorIds:
          type: array
          items: { type: string, format: uuid }
          description: |
            Forma antigua, sin nombre público. Si se manda `monitors`, este
            campo se ignora.
        monitors:
          type: array
          description: Monitores de la página, con su nombre público opcional.
          items: { $ref: "#/components/schemas/StatusPageMonitor" }
    StatusPageMonitor:
      type: object
      required: [id]
      properties:
        id: { type: string, format: uuid }
        displayName:
          type: string
          nullable: true
          maxLength: 200
          description: |
            Nombre con el que se enseña el monitor en la página pública.
            null o ausente = se usa el nombre del monitor.
    StatusPage:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        title: { type: string }
        isPublic: { type: boolean }
        monitorIds:
          type: array
          items: { type: string, format: uuid }
          description: Solo en POST/PATCH/GET de una página concreta, no en el listado.
        monitors:
          type: array
          items: { $ref: "#/components/schemas/StatusPageMonitor" }
          description: Solo en POST/PATCH/GET de una página concreta, no en el listado.
        createdAt: { type: string, format: date-time }
    PublicStatus:
      type: object
      properties:
        title: { type: string }
        overallStatus: { type: string, enum: [operational, degraded, major_outage] }
        monitors:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              name: { type: string }
              currentStatus: { type: string, enum: [up, degraded, down, paused, pending] }
              uptime24h: { type: number, nullable: true }
              uptime7d: { type: number, nullable: true }
              uptime30d: { type: number, nullable: true }
              incidents:
                type: array
                items: { $ref: "#/components/schemas/Incident" }
  responses:
...
    TooMany:
      description: Rate limit.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
paths:
...
  /status-pages:
    get:
      tags: [Status pages]
      summary: Status pages de la organización
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/organizationHeader" }]
      responses:
        "200": { description: "Lista.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/StatusPage" } } } } }
    post:
      tags: [Status pages]
      summary: Crear (editor). Pública en `/status/{username}/{slug}`.
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/organizationHeader" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/StatusPageInput" } } }
      responses:
        "201": { description: "Creada.", content: { application/json: { schema: { $ref: "#/components/schemas/StatusPage" } } } }
        "409": { description: "Slug repetido en la organización." }
  /status-pages/{id}:
    get:
      tags: [Status pages]
      summary: Detalle
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: "Detalle.", content: { application/json: { schema: { $ref: "#/components/schemas/StatusPage" } } } }
    patch:
      tags: [Status pages]
      summary: Editar título, visibilidad o monitores (editor)
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        content: { application/json: { schema: { $ref: "#/components/schemas/StatusPageInput" } } }
      responses:
        "200": { description: "Actualizada.", content: { application/json: { schema: { $ref: "#/components/schemas/StatusPage" } } } }
    delete:
      tags: [Status pages]
      summary: Borrar (editor)
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "204": { description: "Borrada." }
  /public/status/team/{orgSlug}/{slug}:
    get:
      tags: [Público]
      summary: Status page pública de una organización (sin sesión)
      security: []
      description: |
        Misma respuesta que `/public/status/{username}/{slug}`, pero
        resolviendo por el identificador público de la ORGANIZACIÓN
        (`organizations.slug`) en vez de por el username de una persona. Es la
        única forma de que una organización de equipo tenga URL propia: la
        ruta por username siempre resuelve a la organización personal de ese
        usuario. 404 idéntico si la organización o el slug no existen.
      parameters:
        - { name: orgSlug, in: path, required: true, schema: { type: string } }
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: "Estado.", content: { application/json: { schema: { $ref: "#/components/schemas/PublicStatus" } } } }
        "404": { description: "Página no encontrada." }
        "429": { $ref: "#/components/responses/TooMany" }
  /public/status/{username}/{slug}:
    get:
      tags: [Público]
      summary: Status page pública (sin sesión)
      security: []
      description: 404 idéntico si el usuario o el slug no existen (no permite enumerar usuarios).
      parameters:
        - { name: username, in: path, required: true, schema: { type: string } }
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: "Estado.", content: { application/json: { schema: { $ref: "#/components/schemas/PublicStatus" } } } }
        "404": { description: "Página no encontrada." }
        "429": { $ref: "#/components/responses/TooMany" }